% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountToTPM.R
\name{autoDetectGeneIdType}
\alias{autoDetectGeneIdType}
\title{Detect gene id type}
\usage{
autoDetectGeneIdType(id)
}
\arguments{
\item{id}{Gene id in counts data}
}
\value{
Gene id Type
}
\description{
Detect gene id type
}
