% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondCorrAnal.R
\name{CondCorrAnal}
\alias{CondCorrAnal}
\title{Conditional Correlation Analysis}
\usage{
CondCorrAnal(
  corscore.df,
  raw.count.profile,
  is.rawcount = TRUE,
  gene.length = NULL,
  tpm.exp = NULL,
  mutationCalling.map,
  sig.cutoff = 0.01,
  signature.df,
  case = "1",
  control = "0",
  cor.cutoff = NULL,
  cor.method = "spearman",
  min.sz = 15,
  perm.times = 100,
  pAdjustMethod = "BH"
)
}
\arguments{
\item{corscore.df}{Association between mutations and immunophenotypes calculated by
integrated GSEA algorithms}

\item{raw.count.profile}{Expression profile with row as genes, column as samples.}

\item{tpm.exp}{Given tpm expression profile.}

\item{mutationCalling.map}{A matrix of gene mutation status (0 or 1, in character),
row as genes, column as samples.}

\item{sig.cutoff}{Cutoff for significant associations,sig.cutoff = 0.01 by default.
See the MEIGSEA function for details on the remaining parameters}
}
\value{
Significant correlation pairs
}
\description{
Conditional Correlation Analysis
}
