% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountToTPM.R
\name{CountToTPM}
\alias{CountToTPM}
\title{Normalize counts to gene length}
\usage{
CountToTPM(counts, gene.length = NULL)
}
\arguments{
\item{counts}{Counts data}

\item{lengths}{Genes' lengths. If null use a precalculated gene lengths}
}
\value{
TPM values
}
\description{
Normalize counts to gene length
}
