% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegratedGSEA.R
\name{IntegratedGSEA}
\alias{IntegratedGSEA}
\title{Identifying somatic gene mutations associated with immunophenotype gene sets.}
\usage{
IntegratedGSEA(
  raw.count.profile,
  is.rawcount = TRUE,
  gene.length = NULL,
  tpm.exp = NULL,
  mutationCalling.map,
  signature.df,
  case = "1",
  control = "0",
  cor.cutoff = NULL,
  cor.method = "spearman",
  min.sz = 15,
  perm.times = 100,
  pAdjustMethod = "BH"
)
}
\arguments{
\item{raw.count.profile}{RNA-seq data in raw count or microarray expression profile
with row as genes, column as samples.}

\item{is.rawcount}{whether or not the 'raw.count.profile' is RNA-seq raw count.
For microarray expression data as input, is.rawcount = FALSE.}

\item{gene.length}{Gene length annotation file, the first column is the gene name, the second
column is the gene length, if 'gene.length' is not provided, the 'tpm.exp' needs to be given.}

\item{tpm.exp}{Given tpm expression profile (optional). If the 'gene.lengths' given,
tpm.exp is not needed.}

\item{mutationCalling.map}{A matrix of gene mutation status (0 or 1, in character),
row as genes, column as samples.}

\item{signature.df}{A data frame containing immunophenotype signature: the first column
(named 'Gene') is the gene ID, the second column (named 'setAnno') is the immunophenotype name.}

\item{cor.cutoff}{Setting the cutoff of correlation coefficient when refining the
immunophenotype, cor.cutoff = NULL by default, means all genes with statistical significance
will be retained}

\item{cor.method}{method used to compute the correlation, either "spearman", "pearson"
or "kendall", "spearman" by default}

\item{min.sz}{At least how many immunophenotype genes are matched in the expression profile,
min.sz = 15 by default.}

\item{perm.times}{Perumutation times when the normalization of ssGSEA score is needed,
100 by default.}

\item{pAdjustMethod}{Method used to adjust p values, "BH" by default.}

\item{case/control}{In default, case = "1", control = "0", or the specific character
representing mutant (case) and wild type (control) according to 'mutationCalling.map'.}
}
\value{
A list including detailed corrlation results, refined immunophenotype signature
and all correlated mutation-signature pairs

###############################

###############################

get the detailed results from each sub-method
\subsection{}{
}

\subsection{}{
}

\subsection{}{
}

\subsection{}{
}

\subsection{}{
}

\subsection{}{
}
}
\description{
Identifying somatic gene mutations associated with immunophenotype gene sets.
}
