% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEIGSEA.R
\name{MEIGSEA}
\alias{MEIGSEA}
\title{Main Function for identifying somatic gene mutations associated with
immunophenotype gene sets.}
\usage{
MEIGSEA(
  raw.count.profile,
  is.rawcount = TRUE,
  gene.length = NULL,
  tpm.exp = NULL,
  mutationCalling.map,
  signature.df,
  case = "1",
  control = "0",
  cor.cutoff = NULL,
  cor.method = "spearman",
  min.sz = 15,
  perm.times = 100,
  pAdjustMethod = "BH",
  cond.corr.anal = TRUE,
  sig.cutoff = 0.01,
  output = TRUE,
  output.dir = getwd(),
  interested.mut = NULL
)
}
\arguments{
\item{raw.count.profile}{RNA-seq data in raw count or microarray expression
profile with row as genes, column as samples.}

\item{is.rawcount}{whether or not the 'raw.count.profile' is RNA-seq raw count.
For microarray expression data as input, is.rawcount = FALSE.}

\item{gene.length}{Gene length annotation file, the first column is the gene name,the second
column is the gene length, if 'gene.length' is not provided, the 'tpm.exp' needs to be given.}

\item{tpm.exp}{Given tpm expression profile (optional). If the 'gene.lengths' given,
tpm.exp is not needed.}

\item{mutationCalling.map}{A matrix of gene mutation status (0 or 1, in character),
row as genes, column as samples.}

\item{signature.df}{signature A data frame containing immunophenotype signature:
the first column (named 'Gene') is the gene ID, the second column (named 'setAnno')
is the immunophenotype name.}

\item{cor.cutoff}{Setting the cutoff of correlation coefficient when refining the
immunophenotype, cor.cutoff = NULL by default, means all genes with statistical
significance will be retained}

\item{cor.method}{method used to compute the correlation, either "spearman", "pearson"
or "kendall", "spearman" by default}

\item{min.sz}{At least how many immunophenotype genes are matched in the expression
profile, min.sz = 15 by default.}

\item{perm.times}{Perumutation times when the normalization of ssGSEA score is needed,
100 by default.}

\item{pAdjustMethod}{Method used to adjust p values, "BH" by default.}

\item{cond.corr.anal}{Whether to perform conditional association analysis,
cond.corr.anal=TRUE by default.}

\item{sig.cutoff}{cutoff for significant associations, sig.cutoff = 0.01 by default.}

\item{output}{Whether to store the results locally, output = TRUE by default.}

\item{output.dir}{Character representing the storage path of results.}

\item{interested.mut}{vector of interested gene mutations, for which the detailed
correlation results will be given as independent files}

\item{case/control}{In default, case = "1", control = "0", or the specific character
representing mutant (case) and wild type (control) according to 'mutationCalling.map'.}
}
\value{
a list including detailed corrlation results, refined immunophenotype signature
and all correlated mutation-signature pairs
}
\description{
Main Function for identifying somatic gene mutations associated with
immunophenotype gene sets.
}
