% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDiffAssess.R
\name{scoreDiffAssess}
\alias{scoreDiffAssess}
\title{ssGSEA strategy}
\usage{
scoreDiffAssess(
  exp.profile,
  sig.list,
  min.sz = 15,
  score.norm = FALSE,
  perm.times = 1000,
  is.weightInteg = FALSE,
  sig.weight = NULL,
  sig.group = NULL,
  sample.group,
  case,
  control,
  pAdjustMethod = "BH"
)
}
\arguments{
\item{exp.profile}{Expression profile with row as genes, column as samples.}

\item{sig.list}{Multiple signature gene sets to be enriched.}

\item{min.sz}{At least how many immunophenotype genes are matched in the
expression profile, min.sz = 15 by default.}

\item{score.norm}{Whether ssGSEA scores need to be standardized based on random
background, default is FALSE.}

\item{perm.times}{Permutation times when the normalization of ssGSEA score
is needed, 100 by default.}

\item{is.weightInteg}{Whether or not the ssGSEA scores need to be weighted}

\item{sig.weight}{When is.weightInteg = TRUE, set the weight vector
for the different signatures}

\item{sig.group}{When is.weightInteg = TRUE, set the group to which different signatures belong}

\item{sample.group}{Expression profile sample column label, used to calculate
ssGSEA score differences between samples.}

\item{pAdjustMethod}{Method used to adjust p values, "BH" by default.}

\item{case/control}{In default, case = "1", control = "0", or the specific character
representing mutant (case) and wild type (control) according to 'sample.group'.}
}
\value{
ssGSEA score for each sample and differences in ssGSEA scores between mutant and wild type
\enumerate{
\item Calculate the ssGSEA score for each sample
}
\subsection{}{
}

\subsection{}{
}

\subsection{}{
}

\subsection{}{
\enumerate{
\item If necessary, combine the different signature scores weighted
\item Compare the ssGSEA score between mutant and WT samples using Wilcoxon rank sum test
}
}
}
\description{
ssGSEA strategy
}
