% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailedGSEA.R
\name{tailedGSEA}
\alias{tailedGSEA}
\title{one/two-tail GSEA strategy}
\usage{
tailedGSEA(
  exp.profile,
  is.rawcount = TRUE,
  sample.group,
  case,
  control,
  sig.df,
  min.sz = 15,
  pvalueCutoff = 1,
  pAdjustMethod = "BH",
  is.fgsea = FALSE,
  is.fisherPcombine = FALSE
)
}
\arguments{
\item{exp.profile}{RNA-seq data in raw count or microarray expression profile
with row as genes, column as samples.}

\item{is.rawcount}{Whether or not the 'raw.count.profile' is RNA-seq raw count.
For microarray expression data as input, is.rawcount = FALSE.}

\item{sample.group}{Expression profile sample column label, used to calculate the
differentially expressed log2FC between samples}

\item{sig.df}{Multiple signature gene sets to be enriched.}

\item{min.sz}{At least how many immunophenotype genes are matched in the expression
profile, min.sz = 15 by default.}

\item{pvalueCutoff}{The threshold of GSEA p value, if set as 1, all enrichment
results will be output}

\item{pAdjustMethod}{Method used to adjust p values, "BH" by default.}

\item{is.fgsea}{Whether to call fgsea function directly, default is FALSE.}

\item{is.fisherPcombine}{Whether to integrate p-values from two-tail GSEA,default is FALSE}

\item{case/control}{In default, case = "1", control = "0", or the specific character
representing mutant (case) and wild type (control) according to 'sample.group'.}
}
\value{
Enrichment results of two GSEA strategies, and log2FC
\enumerate{
\item Calculate log2FC  of gene expression differences between MUT and WT samples
\item One-tail GSEA
\item Two-tail GSEA
}
}
\description{
one/two-tail GSEA strategy
}
